feat(statics): Add destruction for optre statics
