--- addons/patches/optre_weapons_staticturret/CfgEventHandlers.inc
@@ -1,7 +1,0 @@
-class Extended_Killed_EventHandlers {
-  class OPTRE_Static_M41 {
-    class SUBADDON {
-      clientKilled = "[{ deleteVehicle _this }, _this select 0, 5] call CBA_fnc_waitAndExecute";
-    };
-  };
-};
