--- addons/patches/optre_weapons_staticturret/script_component.hpp
@@ -1,3 +1,0 @@
-#define SUBCOMPONENT optre_weapons_staticturret
-#define SUBCOMPONENT_BEAUTIFIED Operation: TREBUCHET Weapons - Static Turreet
-#include "..\script_component.hpp"
