+++ addons/statics/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\sws\addons\statics
