+++ addons/statics/XEH_PREP.inc.sqf
@@ -1,0 +1,1 @@
+PREP(optreReplaceStatic);
