+++ addons/statics/XEH_postInit.sqf
@@ -1,0 +1,1 @@
+#include "script_component.hpp"
