+++ addons/statics/XEH_preInit.sqf
@@ -1,0 +1,4 @@
+#include "script_component.hpp"
+ADDON = false;
+#include "XEH_PREP.inc.sqf"
+ADDON = true;
