+++ addons/statics/functions/script_component.inc.sqf
@@ -1,0 +1,1 @@
+#include "..\script_component.hpp"
