+++ addons/statics/script_component.hpp
@@ -1,0 +1,15 @@
+#define COMPONENT statics
+#define COMPONENT_BEAUTIFIED Static Weaponry
+#include "\x\sws\addons\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+
+#ifdef DEBUG_ENABLED_INGAME
+  #define DEBUG_MODE_FULL
+#endif
+#ifdef DEBUG_SETTINGS_STATICS
+  #define DEBUG_SETTINGS DEBUG_SETTINGS_INGAME
+#endif
+
+#include "\x\sws\addons\main\script_macros.hpp"
